# src/tag/tag.router.ts
import express from "express";
import type { Request, Response } from "express";
import { body, validationResult } from "express-validator";
import { token } from "../utils/token.server";
import { tenantMiddleware } from "../utils/tenant.middleware";
import { buildPaginatedResponse, getPaginationParams } from "../utils/pagination";

import * as TagService from "./tag.service";
import * as TagValueService from "./tagValue.service";
import { Tag, TagValue } from "@prisma/client";

export const tagRouter = express.Router();

tagRouter.get("/project/:projectId", token.authMiddleware, async (request: Request, response: Response) => {
    // #swagger.tags = ['Tags']
    // #swagger.description = 'Lista tags de um projeto (inclui tags publicas da org).'
    const projectId: number = parseInt(request.params.projectId);
    try {
        //@todo adiconar validações para ver se usuário está no projeto (gerente ou testador)
        const pagination = getPaginationParams(request.query);
        const result = await TagService.findByPaged({
            OR: [{ projectId }, { projectId: null }],
        }, pagination);
        return response.status(200).json(
            buildPaginatedResponse(result.data, result.total, pagination.page, pagination.limit)
        );
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});

tagRouter.get("/:id", token.authMiddleware, async (request: Request, response: Response) => {
    // #swagger.tags = ['Tags']
    // #swagger.description = 'Busca uma tag por id.'
    const id: number = parseInt(request.params.id);
    try {
        //@todo adiconar validações para ver se usuário está no projeto (gerente ou testador) ou se o tags é pública
        const tag = await TagService.find(id);
        if (tag) {
            return response.status(200).json(tag);
        }
        return response.status(404).json("Tag não encontrada");
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});

tagRouter.post("/:projectId", token.authMiddleware, tenantMiddleware, body("name").isString(), async (request: Request, response: Response) => {
    // #swagger.tags = ['Tags']
    // #swagger.description = 'Cria uma tag para um projeto.'
    const errors = validationResult(request);
    if (!errors.isEmpty()) {
        return response.status(400).json({ errors: errors.array() });
    }
    try {
        const projectId: number = parseInt(request.params.projectId);
        if (!projectId) {
            return response.status(404).json({ error: "Projeto não definido" });
        }
        //@todo adiconar validações para ver se usuário está no projeto (gerente apenas)
        const body = request.body
        const tagValues = request.body.tagValues || [];
        delete body.tagValues;
        const tagData = { ...body, situation: TagService.TagSituationEnum.use, projectId: projectId, organizationId: request.organizationId };
        const newTag = await TagService.create(tagData);
        if (newTag) {
            processAddTagValues(tagValues, newTag.id);
        }
        return response.status(201).json(newTag);
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});

async function processAddTagValues(tagValues: TagValue[], newTagId: number) {
    for (const tagValue of tagValues) {
        tagValue.tagId = newTagId
        await TagValueService.create(tagValue);
    }
}

tagRouter.put("/:id", token.authMiddleware, body("name").isString(), body("situation").isNumeric(), async (request: Request, response: Response) => {
    // #swagger.tags = ['Tags']
    // #swagger.description = 'Atualiza dados de uma tag.'
    const id: number = parseInt(request.params.id);
    const errors = validationResult(request);
    if (!errors.isEmpty()) {
        return response.status(400).json({ errors: errors.array() });
    }
    try {
        const tag = await TagService.find(id);
        if (!tag) {
            return response.status(404).json("Tag não encontrada");
        }
        //@todo adiconar validações para ver se usuário está no projeto (gerente apenas)
        const body = request.body
        const tagValues = request.body.tagValues || [];
        delete body.tagValues;
        const updatedTag = await TagService.update(id, body);
        processEditTagValues(tagValues, updatedTag.id);
        return response.status(200).json(updatedTag);
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});

async function processEditTagValues(tagValues: TagValue[],updatedTagId: number) {
    for (const tagValue of tagValues) {
        if (tagValue.id) {
            await TagValueService.update(tagValue.id, tagValue);
        } else {
            tagValue.tagId = updatedTagId;
            await TagValueService.create(tagValue);
        }
    }
}

tagRouter.post("/value/:tagId", token.authMiddleware, body("name").isString(), async (request: Request, response: Response) => {
    // #swagger.tags = ['TagValues']
    // #swagger.description = 'Cria um valor para uma tag.'
    const errors = validationResult(request);
    if (!errors.isEmpty()) {
        return response.status(400).json({ errors: errors.array() });
    }
    try {
        const tagId: number = parseInt(request.params.tagId);
        if (!tagId) {
            return response.status(404).json({ error: "Tag não definida" });
        }
        //@todo adiconar validações para ver se usuário está no projeto (gerente apenas)
        const tagValueData = { ...request.body, situation: TagService.TagSituationEnum.use, tagId: tagId };
        const newTagValue = await TagValueService.create(tagValueData);
        return response.status(201).json(newTagValue);
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});

tagRouter.put("/value/:tagValueid", token.authMiddleware, body("name").isString(), body("situation").isNumeric(), async (request: Request, response: Response) => {
    // #swagger.tags = ['TagValues']
    // #swagger.description = 'Atualiza um valor de tag.'
    const id: number = parseInt(request.params.tagValueid);
    const errors = validationResult(request);
    if (!errors.isEmpty()) {
        return response.status(400).json({ errors: errors.array() });
    }
    try {
        const tagValue = await TagValueService.find(id);
        if (!tagValue) {
            return response.status(404).json("Tag Valor não encontrado");
        }
        //@todo adiconar validações para ver se usuário está no projeto (gerente apenas)
        const updatedTagValue = await TagValueService.update(id, request.body);
        return response.status(200).json(updatedTagValue);
    } catch (error: any) {
        return response.status(500).json(error.message);
    }
});